deployment-stacks/requirements.txt
aws-cdk-lib==2.215.0
constructs>=10.0.0,<11.0.0
aws-cdk-aws-amplify-alpha=2.227.0a0
